backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic>=2.5.2,<3.0.0
python-doctr[torch]==0.7.0
mistralai==0.0.8
python-jose==3.3.0
python-dotenv==1.0.0
